content/canvas/test/webgl/conformance/glsl/misc/glsl-long-variable-names.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>glsl long variable name mapping tests</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../resources/webgl-test.js"> </script>
</head>
<body>
    <canvas id="example" width="50" height="50">
    There is supposed to be an example drawing here, but it's not important.
    </canvas>
    <div id="description"></div>
    <div id="console"></div>
    <script id="vshader" type="x-shader/x-vertex">
        attribute vec4 vPosition0123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456;
        varying float alpha01234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890;
        void main()
        {
            alpha01234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890 = 1.0;
            gl_Position = vPosition0123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456;
        }
    </script>

    <script id="fshader" type="x-shader/x-fragment">
        precision mediump float;
        varying float alpha01234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890;
        uniform float color01234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567[3];
        void main()
        {
            for (int i012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234 = 0; i012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234 < 1; ++i012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234)
            {
                gl_FragColor = vec4(color01234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567[0], color01234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567[1], color01234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567[2], alpha01234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890);
            }
        }
    </script>

    <script>
        function fail(x,y, buf, shouldBe)
        {
            var i = (y*50+x) * 4;
            var reason = "pixel at ("+x+","+y+") is ("+buf[i]+","+buf[i+1]+","+buf[i+2]+","+buf[i+3]+"), should be "+shouldBe;
            testFailed(reason);
        }

        function pass()
        {
            testPassed("drawing is correct");
        }

        if (window.initNonKhronosFramework) {
            window.initNonKhronosFramework(false);
        }

        description("Verify that shader long variable names works fine if they are within 256 characters.");

        gl = initWebGL("example", "vshader", "fshader", [ "vPosition0123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456"], [ 0, 0, 0, 1 ], 1);

        var prog = gl.getParameter(gl.CURRENT_PROGRAM);
        shouldBeNonNull(prog);
        var redLoc = gl.getUniformLocation(prog, "color01234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567[0]");
        shouldBeNonNull(redLoc);
        var greenLoc = gl.getUniformLocation(prog, "color01234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567[1]");
        shouldBeNonNull(greenLoc);
        var blueLoc = gl.getUniformLocation(prog, "color01234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567[2]");
        shouldBeNonNull(blueLoc);

        shouldBe("gl.getProgramParameter(prog, gl.ACTIVE_ATTRIBUTES)", "1");
        var activeAttrib = gl.getActiveAttrib(prog, 0);
        shouldBeNonNull(activeAttrib);
        shouldBe("activeAttrib.size", "1");
        shouldBe("activeAttrib.type", "gl.FLOAT_VEC4");
        shouldBe("activeAttrib.name", "'vPosition0123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456'");

        shouldBe("gl.getProgramParameter(prog, gl.ACTIVE_UNIFORMS)", "1");
        var activeUniform = gl.getActiveUniform(prog, 0);
        shouldBeNonNull(activeUniform);
        shouldBe("activeUniform.size", "3");
        shouldBe("activeUniform.type", "gl.FLOAT");
        shouldBe("activeUniform.name", "'color01234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567[0]'");

        gl.uniform1f(redLoc, 1.0); 
        gl.uniform1f(greenLoc, 0.0);
        gl.uniform1f(blueLoc, 1.0);

        var vertexObject = gl.createBuffer();
        gl.bindBuffer(gl.ARRAY_BUFFER, vertexObject);
        gl.bufferData(gl.ARRAY_BUFFER, new Float32Array([ 0,0.5,0, -0.5,-0.5,0, 0.5,-0.5,0 ]), gl.STATIC_DRAW);
        gl.enableVertexAttribArray(0);
        gl.vertexAttribPointer(0, 3, gl.FLOAT, false, 0, 0);

        gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT);
        gl.drawArrays(gl.TRIANGLES, 0, 3);

        var buf = new Uint8Array(50 * 50 * 4);
        gl.readPixels(0, 0, 50, 50, gl.RGBA, gl.UNSIGNED_BYTE, buf);
        shouldBe("gl.getError()", "gl.NO_ERROR");

        function checkPixels()
        {
            // Test several locations
            // First line should be all black
            for (var i = 0; i < 50; ++i)
                if (buf[i*4] != 0 || buf[i*4+1] != 0 || buf[i*4+2] != 0 || buf[i*4+3] != 255) {
                    fail(i, 0, buf, "(0,0,0,255)");
                    return;
                }

            // Line 15 should be red for at least 10 red pixels starting 20 pixels in
            var offset = (15*50+20) * 4;
            for (var i = 0; i < 10; ++i)
                if (buf[offset+i*4] != 255 || buf[offset+i*4+1] != 0 || buf[offset+i*4+2] != 255 || buf[offset+i*4+3] != 255) {
                    fail(20 + i, 15, buf, "(255,0,255,255)");
                    return;
                }
            // Last line should be all black
            offset = (49*50) * 4;
            for (var i = 0; i < 50; ++i)
                if (buf[offset+i*4] != 0 || buf[offset+i*4+1] != 0 || buf[offset+i*4+2] != 0 || buf[offset+i*4+3] != 255) {
                    fail(i, 49, buf, "(0,0,0,255)");
                    return;
                }

            pass();
        }
        checkPixels();
        successfullyParsed = true;
    </script>
    <script src="../../../resources/js-test-post.js"></script>
</body>
</html>
